requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
asyncio
attrs
beautifulsoup4
bs4
certifi
cffi
charset-normalizer
colorama
contourpy
cycler
defusedxml
dnspython
fake-useragent
fonttools
frozenlist
h11
idna
importlib_resources
ipwhois
kiwisolver
lxml
matplotlib
multidict
numpy
outcome
packaging
pillow
pip
propcache
pycparser
pyparsing
PyQt5
PyQt5-Qt5
PyQt5_sip
PySocks
python-dateutil
python-dotenv
requests
selenium
six
sniffio
sortedcontainers
soupsieve
trio
trio-websocket
typing_extensions
urllib3
watchdog
webdriver-manager
websocket-client
wsproto
yarl
pandas